backend/requirements.txt
fastapi
uvicorn
python-multipart
pymupdf==1.24.2
requests
supabase
python-dotenv
huggingface-hub
